tests but
In zero2prod::run we invoke (and await) HttpServer::run. HttpServer::run returns an instance
of Server - when we call .await it starts listening on the address we specified indefinitely: it will
handle incoming requests as they arrive, but it will never shutdown or “complete” on its own.
This implies that spawn_app never returns and our test logic never gets executed.

We need to run our application as a background task.
tokio::spawn comes quite handy here: tokio::spawn takes a future and hands it over to the runtime
for polling, without waiting for its completion; it therefore runs concurrently with downstream futures and tasks (e.g. our test logic).